rmsn fp16 - 1.17e+07

+++ dorado/basecall/nn/TxModel.cpp
@@ -55,8 +55,8 @@ RMSNormImpl::RMSNormImpl(int hidden_size_) : hidden_size(hidden_size_) {
 }
 
 at::Tensor RMSNormImpl::forward(at::Tensor x) {
-    at::Tensor rstd = torch::rsqrt(x.to(torch::kFloat32).square().mean(-1, true).add_(eps));
-    x.mul_(rstd).mul_(weight).to(torch::kHalf);
+    at::Tensor rstd = torch::rsqrt(x.square().mean(-1, true).add_(eps));
+    x.mul_(rstd).mul_(weight);
     return x;
 }
